+++ index.html
@@ -39,7 +39,6 @@
           <li><a href="/om-meg">Om Meg</a></li>
           <li><a href="/kontakt">Ta Kontakt</a></li>
           <li><a href="/fotogalleri">Fotogalleri</a></li>
-          <li><a href="reviews.html">Anmeldelser</a></li>
         </ul>
       </div>
     </nav>
@@ -177,7 +176,6 @@
             <li><a href="om-meg.html">Om Meg</a></li>
             <li><a href="kontakt.html">Ta Kontakt</a></li>
             <li><a href="fotogalleri.html">Fotogalleri</a></li>
-            <li><a href="reviews.html">Anmeldelser</a></li>
           </ul>
         </div>
         <div class="footer-section">
@@ -201,6 +199,6 @@
     </footer>
 
     <!-- Main JavaScript file -->
-    <script src="script.js" type="module"></script>
+    <script src="script.js"></script>
   </body>
 </html>
