+++ kontakt.html
@@ -34,7 +34,6 @@
           <li><a href="/om-meg">Om Meg</a></li>
           <li><a href="/kontakt" class="active">Ta Kontakt</a></li>
           <li><a href="/fotogalleri">Fotogalleri</a></li>
-          <li><a href="reviews.html">Anmeldelser</a></li>
         </ul>
       </div>
     </nav>
@@ -42,7 +41,7 @@
     <!-- Location Section -->
     <section id="location" class="section">
       <h2>Her finner du oss</h2>
-      <div class="location-info">
+     <div class="location-info">
         <div class="contact-info">
           <h3>Kontaktinformasjon</h3>
           <p>
@@ -98,7 +97,6 @@
             <li><a href="om-meg.html">Om Meg</a></li>
             <li><a href="kontakt.html">Ta Kontakt</a></li>
             <li><a href="fotogalleri.html">Fotogalleri</a></li>
-            <li><a href="reviews.html">Anmeldelser</a></li>
           </ul>
         </div>
         <div class="footer-section">
@@ -122,6 +120,6 @@
     </footer>
 
     <!-- Main JavaScript file -->
-    <script src="script.js" type="module"></script>
+    <script src="script.js"></script>
   </body>
 </html>
