+++ om-meg.html
@@ -14,6 +14,7 @@
         <h1>Om ULF's Bilpleie</h1>
         <p>Velkommen til Ulfs Bilpleie</p>
         <a href="#" class="book-appointment-btn">Bestill Time</a>
+        <div class="scroll-down"></div>
       </div>
     </header>
 
@@ -33,7 +34,6 @@
           <li><a href="/om-meg" class="active">Om Meg</a></li>
           <li><a href="/kontakt">Ta Kontakt</a></li>
           <li><a href="/fotogalleri">Fotogalleri</a></li>
-          <li><a href="reviews.html">Anmeldelser</a></li>
         </ul>
       </div>
     </nav>
@@ -45,7 +45,7 @@
           <h2>Velkommen til ULF's Bilpleie</h2>
         </div>
         <div class="about-image">
-          <img src="images/ulf-smil.png" alt="Ulf's Bilpleie" />
+          <img src="images/ulf-smil-kort2.png" alt="Ulf's Bilpleie" />
         </div>
 
         <div class="about-text">
@@ -86,6 +86,7 @@
             serviceopplevelse hver gang.
           </p>
           <br />
+          <button id="bookAppointment">Bestill Time</button>
         </div>
       </section>
       <button id="back-to-top">⬆</button>
@@ -102,12 +103,10 @@
         <div class="footer-section">
           <h3>Lenker</h3>
           <ul>
-            <li><a href="index.html">Hjem</a></li>
             <li><a href="tjenester.html">Tjenester</a></li>
             <li><a href="om-meg.html">Om Meg</a></li>
             <li><a href="kontakt.html">Ta Kontakt</a></li>
             <li><a href="fotogalleri.html">Fotogalleri</a></li>
-            <li><a href="reviews.html">Anmeldelser</a></li>
           </ul>
         </div>
         <div class="footer-section">
@@ -129,6 +128,6 @@
         <p>&copy; 2025 ULF's Bilpleie. Alle rettigheter forbeholdt.</p>
       </div>
     </footer>
-    <script src="script.js" type="module"></script>
+    <script src="script.js"></script>
   </body>
 </html>
