--- reviews.html
@@ -1,128 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style.css" />
-    <title>Ulfs Bilpleie - Legg til anmeldelse</title>
-    <!-- Firebase SDKs -->
-    <script
-      type="module"
-      src="https://www.gstatic.com/firebasejs/10.8.0/firebase-app.js"
-    ></script>
-    <script
-      type="module"
-      src="https://www.gstatic.com/firebasejs/10.8.0/firebase-database.js"
-    ></script>
-  </head>
-  <!-- Header Section -->
-  <header class="hero">
-    <div class="hero-content">
-      <h1>Ulf's Bilpleie</h1>
-      <p>Velkommen til Ulfs Bilpleie</p>
-      <a href="#" class="book-appointment-btn">Bestill Time</a>
-
-      <!--<button onclick="scrollToSection('services')">Bestill Time</button> -->
-    </div>
-  </header>
-
-  <!-- Navigation Menu -->
-  <nav class="navbar">
-    <div class="navbar-container">
-      <a href="/index" class="logo" id="logo">ULF's Bilpleie</a>
-      <input type="checkbox" id="nav-toggle" class="nav-toggle" />
-      <label for="nav-toggle" class="nav-toggle-label">
-        <span></span>
-        <span></span>
-        <span></span>
-      </label>
-      <ul class="nav-links">
-        <li><a href="index.html">Hjem</a></li>
-        <li><a href="tjenester.html">Tjenester</a></li>
-        <li><a href="om-meg.html">Om Meg</a></li>
-        <li><a href="kontakt.html">Ta Kontakt</a></li>
-        <li><a href="fotogalleri.html">Fotogalleri</a></li>
-        <li><a href="reviews.html" class="active">Anmeldelser</a></li>
-      </ul>
-    </div>
-  </nav>
-  <body class="reviews-page">
-    <br />
-    <br />
-    <div class="container">
-      <h1>Ulfs Bilpleie - Legg til anmeldelse</h1>
-
-      <!-- Input fields for Reviews -->
-      <label for="title">Bilvask:</label>
-      <input type="text" id="title" placeholder="Bilvasken" /><br />
-
-      <label for="rating">Vurdering (1-5)</label>
-      <input
-        type="number"
-        id="rating"
-        min="1"
-        max="5"
-        placeholder="Vurdering fra 1 til 5"
-      /><br />
-
-      <label for="reviewer">Kunde:</label>
-      <input type="text" id="reviewer" placeholder="Ditt navn" /><br />
-
-      <label for="comment">Kommentar</label>
-      <textarea
-        id="comment"
-        placeholder="Skriv anmeldelsen her"
-        rows="4"
-        cols="50"
-      ></textarea>
-      <br />
-
-      <button id="addButton">Legg til anmeldelse</button>
-
-      <h2>Tidligere anmeldelser</h2>
-      <ul id="dataList"></ul>
-    </div>
-    <script type="module" src="script.js"></script>
-  </body>
-  <!-- Back to Top Button -->
-  <button id="back-to-top">⬆</button>
-
-  <!-- Footer Section -->
-  <footer class="footer">
-    <div class="footer-content">
-      <div class="footer-section">
-        <h3>ULF's Bilpleie</h3>
-        <p>Hvittingfossveien 129, 3085, Holmestrand</p>
-        <p>+47 404 98 499</p>
-      </div>
-      <div class="footer-section">
-        <h3>Lenker</h3>
-        <ul>
-          <li><a href="index.html">Hjem</a></li>
-          <li><a href="tjenester.html">Tjenester</a></li>
-          <li><a href="om-meg.html">Om Meg</a></li>
-          <li><a href="kontakt.html">Ta Kontakt</a></li>
-          <li><a href="fotogalleri.html">Fotogalleri</a></li>
-          <li><a href="reviews.html">Anmeldelser</a></li>
-        </ul>
-      </div>
-      <div class="footer-section">
-        <h3>Følg meg på FaceBook</h3>
-        <ul class="social-icons">
-          <li>
-            <a
-              href="https://www.facebook.com/profile.php?id=100063566292197"
-              target="_blank"
-              rel="noopener noreferrer"
-              ><img src="images/fb.png" alt="Facebook"
-            /></a>
-          </li>
-          <span>Ulf's Bilpleie</span>
-        </ul>
-      </div>
-    </div>
-    <div class="footer-bottom">
-      <p>&copy; 2025 ULF's Bilpleie. Alle rettigheter forbeholdt.</p>
-    </div>
-  </footer>
-</html>
