+++ tjenester.html
@@ -36,11 +36,10 @@
         </label>
         <ul class="nav-links">
           <li><a href="/index">Hjem</a></li>
-          <li><a href="/tjenester" class="active">Tjenester </a></li>
+          <li><a href="/tjenester"class="active">Tjenester </a></li>
           <li><a href="/om-meg">Om Meg</a></li>
           <li><a href="/kontakt">Ta Kontakt</a></li>
           <li><a href="/fotogalleri">Fotogalleri</a></li>
-          <li><a href="reviews.html">Anmeldelser</a></li>
         </ul>
       </div>
     </nav>
@@ -104,9 +103,11 @@
           </p>
           <p><strong>Pris fra:</strong> 6500 NOK</p>
         </div>
+        
       </div>
+    
     </section>
-    <section id="services" class="section">
+      <section id="services" class="section">
       <h2>VaskePakker</h2>
 
       <div class="services-grid">
@@ -130,9 +131,11 @@
           </p>
           <p><strong>Pris fra:</strong> 5500 NOK</p>
         </div>
+        
       </div>
+     
     </section>
-    <section id="services" class="section">
+    <section id="services" class="section"> 
       <h2>Tilleggstjenester</h2>
       <div class="extra-services-list">
         <div class="extra-service">
@@ -175,12 +178,10 @@
         <div class="footer-section">
           <h3>Lenker</h3>
           <ul>
-            <li><a href="index.html">Hjem</a></li>
             <li><a href="tjenester.html">Tjenester</a></li>
             <li><a href="om-meg.html">Om Meg</a></li>
             <li><a href="kontakt.html">Ta Kontakt</a></li>
             <li><a href="fotogalleri.html">Fotogalleri</a></li>
-            <li><a href="reviews.html">Anmeldelser</a></li>
           </ul>
         </div>
         <div class="footer-section">
@@ -202,8 +203,8 @@
         <p>&copy; 2025 ULF's Bilpleie. Alle rettigheter forbeholdt.</p>
       </div>
     </footer>
-
+   
     <!-- Main JavaScript file -->
-    <script src="script.js" type="module"></script>
+    <script src="script.js"></script>
   </body>
 </html>
